requirements.txt
yfpy>=5.1.5
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
feedparser>=6.0.0
textblob>=0.19.0
